improve help msg, make copying config files optional

+++ setup.sh
@@ -10,18 +10,25 @@ if [ $? -ne 0 ]; then
 	exit 1
 fi
 
-if [ -d "$HOME/.config/rsp" ]; then
-	mv "$HOME/.config/rsp" "$HOME/.config/rsp_backup"
+read -p "Do you want to copy the configuration files? (y/N) " -n 1 -r
+echo
+
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+	# Backup existing configuration if it exists
+	if [ -d "$HOME/.config/rsp" ]; then
+		mv "$HOME/.config/rsp" "$HOME/.config/rsp_backup"
+		if [ $? -ne 0 ]; then
+			echo "Failed to rename existing 'rsp' directory."
+			exit 1
+		fi
+	fi
+
+	# Copy new configuration files
+	cp -r ./config/rsp/ "$HOME/.config/"
 	if [ $? -ne 0 ]; then
-		echo "Failed to rename existing 'rsp' directory."
+		echo "Failed to copy configuration files."
 		exit 1
 	fi
-fi
-
-cp -r ./config/rsp/ "$HOME/.config/"
-if [ $? -ne 0 ]; then
-	echo "Failed to copy configuration files."
-	exit 1
 fi
 
 echo "Successful installation."
